Moved components out of starter
