+++ src/components/gauge/gauge.module.css
