+++ src/components/header/header.tsx
@@ -1,6 +1,6 @@
 import { component$ } from "@builder.io/qwik";
 import styles from "./header.module.css";
-import upayan from "../../../media/upayan.svg";
+import upayan from "../../media/upayan.svg";
 
 export default component$(() => {
   return (
