+++ src/routes/index.tsx
@@ -1,7 +1,7 @@
 import { component$ } from "@builder.io/qwik";
 import type { DocumentHead } from "@builder.io/qwik-city";
 
-import Hero from "../components/starter/hero/hero";
+import Hero from "../components/hero/hero";
 
 export default component$(() => {
   return (
