+++ src/routes/layout.tsx
@@ -2,8 +2,8 @@ import { component$, Slot, useStyles$ } from "@builder.io/qwik";
 import { routeLoader$ } from "@builder.io/qwik-city";
 import type { RequestHandler } from "@builder.io/qwik-city";
 
-import Header from "../components/starter/header/header";
-import Footer from "../components/starter/footer/footer";
+import Header from "../components/header/header";
+import Footer from "../components/footer/footer";
 
 import styles from "./styles.css?inline";
